insert-only token account

+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/domain/Entity.java
@@ -53,7 +53,7 @@ public class Entity {
 
     private byte[] key;
 
-    private Long maxAutomaticTokenAssociations;
+    private Integer maxAutomaticTokenAssociations;
 
     @JsonSerialize(using = NullableStringSerializer.class)
     private String memo;
